rename: /components/menu-icons → /components/icons
