+++ components/icons/MonitoringIcon.tsx
