+++ components/icons/SalesIcon.tsx
