+++ components/icons/SettingsIcon.tsx
